+++ libraries/liveIntentId/shared.js
