+++ modules/liveIntentIdSystem.js
@@ -4,10 +4,10 @@ function loadModule() {
   // eslint-disable-next-line no-constant-condition
   if ('$$LIVE_INTENT_MODULE_MODE$$' === 'external') {
     // eslint-disable-next-line no-restricted-globals
-    return require('../libraries/liveIntentIdSystem/liveIntentExternalIdSystem.js')
+    return require('../libraries/liveIntentId/externalIdSystem.js')
   } else {
     // eslint-disable-next-line no-restricted-globals
-    return require('../libraries/liveIntentIdSystem/liveIntentIdSystem.js')
+    return require('../libraries/liveIntentId/idSystem.js')
   }
 }
